fastV2/requirements.txt
# FastAPI RAG Backend Dependencies - Production Grade

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI & Embeddings
openai>=1.0.0

# Vector Store - LanceDB (Disk-based, scalable)
lancedb>=0.5.0
pyarrow>=14.0.0
pandas>=2.0.0

# Background Task Processing
celery>=5.3.0
redis>=5.0.0

# PDF Processing (unstructured.io)
unstructured[pdf]>=0.10.0
pdf2image>=1.16.0
pytesseract>=0.3.10
pillow>=10.0.0

# Environment
python-dotenv>=1.0.0

# Note: Create a .env file with your OpenAI API key
# OPENAI_API_KEY=your_api_key_here

# Production Deployment Notes:
# 1. Install Redis: https://redis.io/docs/getting-started/
#    Windows: https://github.com/microsoftarchive/redis/releases
#    Linux/Mac: brew install redis or apt-get install redis
#
# 2. Start Redis: redis-server
#
# 3. Start Celery Worker:
#    Windows: celery -A celery_worker worker --loglevel=info --pool=solo
#    Linux/Mac: celery -A celery_worker worker --loglevel=info
#
# 4. Start FastAPI: uvicorn main:app --reload --port 8000
#celery poopelr tesseract